refactor(shell): 拆分壳层并明确系统字体能力
- 按职责拆分桌面、移动、状态横幅、更多面板与队列覆盖层
- 用 AppBranch 语义集合保持画廊与素材库生命周期
- 仅在 Windows 枚举系统字体，并让原生桥错误显式传播
- 补充平台能力、字体链路与 Shell 等价性回归测试
